requirements.txt
fastapi
uvicorn
instaloader
python-docx
python-multipart
python-dotenv
pydantic
